# Conflicts:
#	musteri-service/src/main/java/com/marul/musteri/MusteriService.java
#	satis-service/pom.xml

+++ musteri-service/src/main/java/com/marul/musteri/MusteriRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MusteriRepository extends JpaRepository<Musteri, Integer> {
+public interface MusteriRepository extends JpaRepository<Musteri, Long> {
     boolean existsByEmail(String email);
 }
