+++ satis-service/src/main/java/com/marul/satis/SatisService.java
@@ -2,10 +2,6 @@ package com.marul.satis;
 
 import com.marul.dto.MusteriDto;
 import com.marul.dto.SatisDto;
-import com.marul.dto.musteri.MusteriResponseDto;
-import com.marul.dto.result.DataResult;
-import com.marul.dto.result.Result;
-import com.marul.dto.result.SuccessDataResult;
 import com.marul.util.ResultDecoder;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -29,7 +25,7 @@ public class SatisService {
 
     private SatisDto getSatisDto(Satis satis) {
         Long musteriId = satis.getMusteriId();
-        MusteriDto musteriDto = ResultDecoder.getDataResult(musteriFeignClient.findById(musteriId.intValue()));
+        MusteriDto musteriDto = ResultDecoder.getDataResult(musteriFeignClient.findById(musteriId));
         SatisDto satisDto = new SatisDto();
         satisDto.setMusteriDto(musteriDto);
         return satisDto;
@@ -37,7 +33,7 @@ public class SatisService {
 
     public SatisDto save(SatisDto satisDto) {
         Satis satis = satisMapper.getEntity(satisDto);
-        ResultDecoder.getDataResult(musteriFeignClient.findById(satis.getMusteriId().intValue()));
+        ResultDecoder.getDataResult(musteriFeignClient.findById(satis.getMusteriId()));
         satis = this.satisRepository.save(satis);
         return satisMapper.getDto(satis);
     }
